+++ modules/alloy/src/main/ts/ephox/alloy/positioning/view/Bounder.ts
@@ -1,6 +1,6 @@
 import { Adt, Arr, Fun } from '@ephox/katamari';
-
 import { Bounds } from '../../alien/Boxes';
+import { cap } from '../../alien/Cycles';
 import { Bubble } from '../layout/Bubble';
 import * as Direction from '../layout/Direction';
 import { AnchorBox, AnchorElement, AnchorLayout } from '../layout/LayoutTypes';
@@ -18,20 +18,11 @@ const adt: {
   { nofit: [ 'reposition', 'deltaW', 'deltaH' ] }
 ]);
 
-const attempt = (candidate: SpotInfo, width: number, height: number, bounds: Bounds): BounderAttemptAdt  => {
-  const candidateX = candidate.x();
-  const candidateY = candidate.y();
-  const bubbleLeft = candidate.bubble().offset().left();
-  const bubbleTop = candidate.bubble().offset().top();
-
+const calcReposition = (newX: number, newY: number, width: number, height: number, bounds: Bounds) => {
   const boundsX = bounds.x();
   const boundsY = bounds.y();
   const boundsWidth = bounds.width();
   const boundsHeight = bounds.height();
-
-  // candidate position is excluding the bubble, so add those values as well
-  const newX = candidateX + bubbleLeft;
-  const newY = candidateY + bubbleTop;
 
   // simple checks for "is the top left inside the view"
   const xInBounds = newX >= boundsX;
@@ -44,19 +35,43 @@ const attempt = (candidate: SpotInfo, width: number, height: number, bounds: Bou
   const sizeInBounds = xFit && yFit;
 
   // measure how much of the width and height are visible. deltaW isn't necessary in the fit case but it's cleaner to read here.
-  const deltaW = xInBounds ? Math.min(width, boundsX + boundsWidth - newX)
-                         : Math.abs(boundsX - (newX + width));
-  const deltaH = yInBounds ? Math.min(height, boundsY + boundsHeight - newY)
-                         : Math.abs(boundsY - (newY + height));
+  const deltaW = Math.abs(Math.min(width, xInBounds ? boundsX + boundsWidth - newX : boundsX - (newX + width)));
+  const deltaH = Math.abs(Math.min(height, yInBounds ? boundsY + boundsHeight - newY : boundsY - (newY + height)));
 
   // TBIO-3366 + TBIO-4236:
   // Futz with the X position to ensure that x is positive, but not off the right side of the screen.
-  const maxX = bounds.x() + bounds.width();
-  const minX = Math.max(bounds.x(), newX);
-  const limitX = Math.min(minX, maxX);
-
+  // NOTE: bounds.x() is 0 in repartee here.
+  const limitX = cap(newX, bounds.x(), bounds.right());
   // Futz with the Y value to ensure that we're not off the top of the screen
-  const limitY = yInBounds ? newY : newY + (height - deltaH);
+  const limitY = cap(newY, bounds.y(), bounds.bottom());
+
+  return {
+    originInBounds,
+    sizeInBounds,
+    limitX,
+    limitY,
+    deltaW,
+    deltaH
+  };
+};
+
+const attempt = (candidate: SpotInfo, width: number, height: number, bounds: Bounds): BounderAttemptAdt  => {
+  const candidateX = candidate.x();
+  const candidateY = candidate.y();
+  const bubbleLeft = candidate.bubble().offset().left();
+  const bubbleTop = candidate.bubble().offset().top();
+
+  const boundsY = bounds.y();
+  const boundsHeight = bounds.height();
+
+  const boundsX = bounds.x();
+  const boundsWidth = bounds.width();
+
+  // candidate position is excluding the bubble, so add those values as well
+  const newX = candidateX + bubbleLeft;
+  const newY = candidateY + bubbleTop;
+
+  const { originInBounds, sizeInBounds, limitX, limitY, deltaW, deltaH } = calcReposition(newX, newY, width, height, bounds);
 
   // TBIO-3367 + TBIO-3387:
   // Futz with the "height" of the popup to ensure if it doesn't fit it's capped at the available height.
@@ -65,7 +80,10 @@ const attempt = (candidate: SpotInfo, width: number, height: number, bounds: Bou
   const downAvailable = Fun.constant((boundsY + boundsHeight) - limitY);
   const maxHeight = Direction.cataVertical(candidate.direction(), downAvailable, /* middle */ downAvailable, upAvailable);
 
-  // We don't futz with the width.
+  const westAvailable = Fun.constant((limitX + deltaW) - boundsX);
+  const eastAvailable = Fun.constant((boundsX + boundsWidth) - limitX);
+  const bestAvailable = Fun.constant(Math.max(westAvailable(), eastAvailable()));
+  const maxWidth = Direction.cataHorizontal(candidate.direction(), eastAvailable, /* middle */ bestAvailable, westAvailable);
 
   const reposition = Reposition.decision({
     x: limitX,
@@ -73,6 +91,7 @@ const attempt = (candidate: SpotInfo, width: number, height: number, bounds: Bou
     width: deltaW,
     height: deltaH,
     maxHeight,
+    maxWidth,
     direction: candidate.direction(),
     classes: {
       on: candidate.bubble().classesOn(),
@@ -151,6 +170,7 @@ const attempts = (candidates: AnchorLayout[], anchorBox: AnchorBox, elementBox: 
       width: elementBox.width(),
       height: elementBox.height(),
       maxHeight: elementBox.height(),
+      maxWidth: elementBox.width(),
       direction: Direction.southeast(),
       classes: {
         on: [],
@@ -166,6 +186,4 @@ const attempts = (candidates: AnchorLayout[], anchorBox: AnchorBox, elementBox: 
   return abc.fold(Fun.identity, Fun.identity) as Reposition.RepositionDecision;
 };
 
-export {
-  attempts
-};
+export { attempts, calcReposition };