+++ modules/tinymce/src/core/main/ts/api/UndoManager.ts
@@ -6,9 +6,11 @@
  */
 
 import { Event } from '@ephox/dom-globals';
-import GetBookmark from '../bookmark/GetBookmark';
-import Levels, { UndoLevel } from '../undo/Levels';
-import Tools from './util/Tools';
+import { Cell, Option } from '@ephox/katamari';
+import { Bookmark } from '../bookmark/BookmarkTypes';
+import { addUndoLevel, beforeChange, clear, extra, hasRedo, hasUndo, ignore, redo, reset, transact, undo } from '../undo/Operations';
+import { addKeyboardShortcuts, registerEvents } from '../undo/Setup';
+import { UndoManager, Locks, Index, UndoLevel } from '../undo/UndoManagerTypes';
 import Editor from './Editor';
 
 /**
@@ -16,174 +18,14 @@ import Editor from './Editor';
  *
  * @class tinymce.UndoManager
  */
-
-interface UndoManager {
-  data: UndoLevel[];
-  typing: boolean;
-  add: (level?: UndoLevel, event?: Event) => UndoLevel;
-  beforeChange: () => void;
-  undo: () => UndoLevel;
-  redo: () => UndoLevel;
-  clear: () => void;
-  reset: () => void;
-  hasUndo: () => boolean;
-  hasRedo: () => boolean;
-  transact: (callback: () => void) => UndoLevel;
-  ignore: (callback: () => void) => void;
-  extra: (callback1: () => void, callback2: () => void) => void;
-}
-
 const UndoManager = function (editor: Editor): UndoManager {
-  let self: UndoManager = this, index = 0, data = [], beforeBookmark, isFirstTypedCharacter, locks = 0;
-
-  const isUnlocked = function () {
-    return locks === 0;
-  };
-
-  const setTyping = function (typing) {
-    if (isUnlocked()) {
-      self.typing = typing;
-    }
-  };
-
-  const setDirty = function (state) {
-    editor.setDirty(state);
-  };
-
-  const addNonTypingUndoLevel = function (e?) {
-    setTyping(false);
-    self.add({} as UndoLevel, e);
-  };
-
-  const endTyping = function () {
-    if (self.typing) {
-      setTyping(false);
-      self.add();
-    }
-  };
-
-  // Add initial undo level when the editor is initialized
-  editor.on('init', function () {
-    self.add();
-  });
-
-  // Get position before an execCommand is processed
-  editor.on('BeforeExecCommand', function (e) {
-    const cmd = e.command;
-
-    if (cmd !== 'Undo' && cmd !== 'Redo' && cmd !== 'mceRepaint') {
-      endTyping();
-      self.beforeChange();
-    }
-  });
-
-  // Add undo level after an execCommand call was made
-  editor.on('ExecCommand', function (e) {
-    const cmd = e.command;
-
-    if (cmd !== 'Undo' && cmd !== 'Redo' && cmd !== 'mceRepaint') {
-      addNonTypingUndoLevel(e);
-    }
-  });
-
-  editor.on('ObjectResizeStart cut', function () {
-    self.beforeChange();
-  });
-
-  editor.on('SaveContent ObjectResized blur', addNonTypingUndoLevel);
-  editor.on('dragend', addNonTypingUndoLevel);
-
-  editor.on('keyup', function (e) {
-    const keyCode = e.keyCode;
-
-    // If key is prevented then don't add undo level
-    // This would happen on keyboard shortcuts for example
-    if (e.isDefaultPrevented()) {
-      return;
-    }
-
-    if ((keyCode >= 33 && keyCode <= 36) || (keyCode >= 37 && keyCode <= 40) || keyCode === 45 || e.ctrlKey) {
-      addNonTypingUndoLevel();
-      editor.nodeChanged();
-    }
-
-    if (keyCode === 46 || keyCode === 8) {
-      editor.nodeChanged();
-    }
-
-    // Fire a TypingUndo/Change event on the first character entered
-    if (isFirstTypedCharacter && self.typing && Levels.isEq(Levels.createFromEditor(editor), data[0]) === false) {
-      if (editor.isDirty() === false) {
-        setDirty(true);
-        editor.fire('change', { level: data[0], lastLevel: null });
-      }
-
-      editor.fire('TypingUndo');
-      isFirstTypedCharacter = false;
-      editor.nodeChanged();
-    }
-  });
-
-  editor.on('keydown', function (e) {
-    const keyCode = e.keyCode;
-
-    // If key is prevented then don't add undo level
-    // This would happen on keyboard shortcuts for example
-    if (e.isDefaultPrevented()) {
-      return;
-    }
-
-    // Is character position keys left,right,up,down,home,end,pgdown,pgup,enter
-    if ((keyCode >= 33 && keyCode <= 36) || (keyCode >= 37 && keyCode <= 40) || keyCode === 45) {
-      if (self.typing) {
-        addNonTypingUndoLevel(e);
-      }
-
-      return;
-    }
-
-    // If key isn't Ctrl+Alt/AltGr
-    const modKey = (e.ctrlKey && !e.altKey) || e.metaKey;
-    if ((keyCode < 16 || keyCode > 20) && keyCode !== 224 && keyCode !== 91 && !self.typing && !modKey) {
-      self.beforeChange();
-      setTyping(true);
-      self.add({} as UndoLevel, e);
-      isFirstTypedCharacter = true;
-    }
-  });
-
-  editor.on('mousedown', function (e) {
-    if (self.typing) {
-      addNonTypingUndoLevel(e);
-    }
-  });
-
-  // Special inputType, currently only Chrome implements this: https://www.w3.org/TR/input-events-2/#x5.1.2-attributes
-  const isInsertReplacementText = (event) => event.inputType === 'insertReplacementText';
-  // Safari just shows inputType `insertText` but with data set to null so we can use that
-  const isInsertTextDataNull = (event) => event.inputType === 'insertText' && event.data === null;
-
-  // For detecting when user has replaced text using the browser built-in spell checker
-  editor.on('input', (e) => {
-    if (e.inputType && (isInsertReplacementText(e) || isInsertTextDataNull(e))) {
-      addNonTypingUndoLevel(e);
-    }
-  });
-
-  // Add keyboard shortcuts for undo/redo keys
-  editor.addShortcut('meta+z', '', 'Undo');
-  editor.addShortcut('meta+y,meta+shift+z', '', 'Redo');
-
-  editor.on('AddUndo Undo Redo ClearUndos', function (e) {
-    if (!e.isDefaultPrevented()) {
-      editor.nodeChanged();
-    }
-  });
+  const beforeBookmark: Cell<Option<Bookmark>> = Cell(Option.none());
+  const locks: Locks = Cell(0);
+  const index: Index = Cell(0);
 
   /*eslint consistent-this:0 */
-  self = {
-    // Explode for debugging reasons
-    data,
+  const undoManager = {
+    data: [], // Gets mutated both internally and externally by plugins like remark, not documented
 
     /**
      * State if the user is currently typing or not. This will add a typing operation into one undo
@@ -200,9 +42,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @method beforeChange
      */
     beforeChange () {
-      if (isUnlocked()) {
-        beforeBookmark = GetBookmark.getUndoBookmark(editor.selection);
-      }
+      beforeChange(editor, locks, beforeBookmark);
     },
 
     /**
@@ -214,66 +54,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Object} Undo level that got added or null it a level wasn't needed.
      */
     add (level?: UndoLevel, event?: Event): UndoLevel {
-      let i;
-      const settings = editor.settings;
-      let lastLevel, currentLevel;
-
-      currentLevel = Levels.createFromEditor(editor);
-      level = level || {} as UndoLevel;
-      level = Tools.extend(level, currentLevel);
-
-      if (isUnlocked() === false || editor.removed) {
-        return null;
-      }
-
-      lastLevel = data[index];
-      if (editor.fire('BeforeAddUndo', { level, lastLevel, originalEvent: event }).isDefaultPrevented()) {
-        return null;
-      }
-
-      // Add undo level if needed
-      if (lastLevel && Levels.isEq(lastLevel, level)) {
-        return null;
-      }
-
-      // Set before bookmark on previous level
-      if (data[index]) {
-        data[index].beforeBookmark = beforeBookmark;
-      }
-
-      // Time to compress
-      if (settings.custom_undo_redo_levels) {
-        if (data.length > settings.custom_undo_redo_levels) {
-          for (i = 0; i < data.length - 1; i++) {
-            data[i] = data[i + 1];
-          }
-
-          data.length--;
-          index = data.length;
-        }
-      }
-
-      // Get a non intrusive normalized bookmark
-      level.bookmark = GetBookmark.getUndoBookmark(editor.selection);
-
-      // Crop array if needed
-      if (index < data.length - 1) {
-        data.length = index + 1;
-      }
-
-      data.push(level);
-      index = data.length - 1;
-
-      const args = { level, lastLevel, originalEvent: event };
-
-      editor.fire('AddUndo', args);
-
-      if (index > 0) {
-        setDirty(true);
-        editor.fire('change', args);
-      }
-
-      return level;
+      return addUndoLevel(editor, undoManager, index, locks, beforeBookmark, level, event);
     },
 
     /**
@@ -283,22 +64,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Object} Undo level or null if no undo was performed.
      */
     undo (): UndoLevel {
-      let level: UndoLevel;
-
-      if (self.typing) {
-        self.add();
-        self.typing = false;
-        setTyping(false);
-      }
-
-      if (index > 0) {
-        level = data[--index];
-        Levels.applyToEditor(editor, level, true);
-        setDirty(true);
-        editor.fire('Undo', { level });
-      }
-
-      return level;
+      return undo(editor, undoManager, locks, index);
     },
 
     /**
@@ -308,16 +74,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Object} Redo level or null if no redo was performed.
      */
     redo (): UndoLevel {
-      let level: UndoLevel;
-
-      if (index < data.length - 1) {
-        level = data[++index];
-        Levels.applyToEditor(editor, level, false);
-        setDirty(true);
-        editor.fire('Redo', { level });
-      }
-
-      return level;
+      return redo(editor, index, undoManager.data);
     },
 
     /**
@@ -326,11 +83,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @method clear
      */
     clear () {
-      data = [];
-      index = 0;
-      self.typing = false;
-      self.data = data;
-      editor.fire('ClearUndos');
+      clear(editor, undoManager, index);
     },
 
     /**
@@ -339,8 +92,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @method reset
      */
     reset () {
-      self.clear();
-      self.add();
+      reset(undoManager);
     },
 
     /**
@@ -350,8 +102,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Boolean} true/false if the undo manager has any undo levels.
      */
     hasUndo () {
-      // Has undo levels or typing and content isn't the same as the initial level
-      return index > 0 || (self.typing && data[0] && !Levels.isEq(Levels.createFromEditor(editor), data[0]));
+      return hasUndo(editor, undoManager, index);
     },
 
     /**
@@ -361,7 +112,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Boolean} true/false if the undo manager has any redo levels.
      */
     hasRedo () {
-      return index < data.length - 1 && !self.typing;
+      return hasRedo(undoManager, index);
     },
 
     /**
@@ -375,10 +126,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @return {Object} Undo level that got added or null it a level wasn't needed.
      */
     transact (callback: () => void): UndoLevel {
-      endTyping();
-      self.beforeChange();
-      self.ignore(callback);
-      return self.add();
+      return transact(undoManager, locks, callback);
     },
 
     /**
@@ -390,12 +138,7 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @param {function} callback Function that gets executed and has dom manipulation logic in it.
      */
     ignore (callback: () => void) {
-      try {
-        locks++;
-        callback();
-      } finally {
-        locks--;
-      }
+      ignore(locks, callback);
     },
 
     /**
@@ -408,21 +151,14 @@ const UndoManager = function (editor: Editor): UndoManager {
      * @param {function} callback2 Function that does mutation but gets displayed to the user.
      */
     extra (callback1: () => void, callback2: () => void) {
-      let lastLevel, bookmark;
-
-      if (self.transact(callback1)) {
-        bookmark = data[index].bookmark;
-        lastLevel = data[index - 1];
-        Levels.applyToEditor(editor, lastLevel, true);
-
-        if (self.transact(callback2)) {
-          data[index - 1].beforeBookmark = bookmark;
-        }
-      }
+      extra(editor, undoManager, index, callback1, callback2);
     }
   };
 
-  return self;
+  registerEvents(editor, undoManager, locks);
+  addKeyboardShortcuts(editor);
+
+  return undoManager;
 };
 
 export default UndoManager;