+++ modules/tinymce/src/plugins/table/main/ts/ui/TableDialogGeneralTab.ts
@@ -8,12 +8,14 @@ const getItems = (editor: Editor, hasClasses: boolean, insertNewTable: boolean) 
     {
       type: 'input',
       name: 'cols',
-      label: 'Cols'
+      label: 'Cols',
+      inputMode: 'numeric'
     },
     {
       type: 'input',
       name: 'rows',
-      label: 'Rows'
+      label: 'Rows',
+      inputMode: 'numeric'
     }
   ];
 
@@ -34,17 +36,20 @@ const getItems = (editor: Editor, hasClasses: boolean, insertNewTable: boolean) 
     {
       type: 'input',
       name: 'cellspacing',
-      label: 'Cell spacing'
+      label: 'Cell spacing',
+      inputMode: 'numeric'
     },
     {
       type: 'input',
       name: 'cellpadding',
-      label: 'Cell padding'
+      label: 'Cell padding',
+      inputMode: 'numeric'
     },
     {
       type: 'input',
       name: 'border',
-      label: 'Border width'
+      label: 'Border width',
+      inputMode: 'numeric'
     },
     {
       type: 'label',