+++ modules/tinymce/src/themes/silver/main/ts/ui/toolbar/button/ToolbarButtons.ts
@@ -16,7 +16,6 @@ import {
   Focusing,
   Keying,
   NativeEvents,
-  Receiving,
   Reflecting,
   Replacing,
   SketchSpec,
@@ -25,6 +24,7 @@ import {
   SystemEvents,
   TieredData,
   TieredMenuTypes,
+  Unselecting,
 } from '@ephox/alloy';
 import { Toolbar, Types } from '@ephox/bridge';
 import { Arr, Cell, Fun, Future, Id, Merger, Option } from '@ephox/katamari';
@@ -32,7 +32,6 @@ import { Attr, SelectorFind } from '@ephox/sugar';
 
 import I18n from 'tinymce/core/api/util/I18n';
 import { UiFactoryBackstageProviders, UiFactoryBackstageShared } from 'tinymce/themes/silver/backstage/Backstage';
-import * as Channels from '../../../Channels';
 import { DisablingConfigs } from '../../alien/DisablingConfigs';
 import { detectSize } from '../../alien/FlatgridAutodetect';
 import { SimpleBehaviours } from '../../alien/SimpleBehaviours';
@@ -322,20 +321,12 @@ const renderSplitButton = (spec: Toolbar.ToolbarSplitButton, sharedBackstage: Ui
 
     splitDropdownBehaviours: Behaviour.derive([
       DisablingConfigs.splitButton(false),
-      Receiving.config({
-        channels: {
-          [ Channels.reposition() ]: {
-            onReceive: (comp) => {
-              AlloySplitDropdown.repositionMenus(comp);
-            }
-          }
-        }
-      }),
       AddEventsBehaviour.config('split-dropdown-events', [
         AlloyEvents.run(focusButtonEvent, Focusing.focus),
         onControlAttached(specialisation, editorOffCell),
         onControlDetached(specialisation, editorOffCell),
-      ])
+      ]),
+      Unselecting.config({ })
     ]),
 
     eventOrder: {